+++ R/generate_transitive_closure.R
@@ -81,7 +81,7 @@ generate_transitive_closure <- function(df, cl = NULL) {
 #'
 #' @param mat A matrix in which column 1 is the lower ranked item and column 2 is the
 #'   upper ranked item.
-#' @keywords internal
+#' @noRd
 .generate_transitive_closure <- function(mat) {
   # This line was an answer to StackOverflow question 51794127
   my_set <- do.call(sets::set, apply(mat, 1, sets::as.tuple))