+++ src/smc_get_exponent_sum.cpp
@@ -22,9 +22,8 @@ using namespace arma;
 //' @return Exponent in the Mallows log likelihood. Note that it does not include
 //' the partition function, and since the partition function depends on \code{alpha},
 //' this is not a likelihood per se.
-//' @export
 //' @author Anja Stein
-//' @keywords internal
+//' @noRd
 //' @examples
 //' set.seed(101)
 //' rho <- t(c(1, 2, 3, 4, 5, 6))